requirements.txt
azure-common==1.1.28
azure-core==1.26.1
azure-identity==1.12.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.6.0
azure-keyvault-keys==4.7.0
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.2
azure-mgmt-resource==21.1.0
azure-storage-blob==12.14.1
scikit-learn==1.0.2
scikit-image==0.19.2
scipy==1.7.3
cython==0.29.28
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
idna==3.4
isodate==0.6.1
msal==1.20.0
msal-extensions==1.0.0
msrest==0.7.1
oauthlib==3.2.2
Office365-REST-Python-Client==2.3.16
openpyxl==3.0.10
pandas==1.5.2
portalocker==2.6.0
pycparser==2.21
PyJWT==2.6.0
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
SQLAlchemy==2.0.0
urllib3==1.26.13
statsmodels==0.13.2
google-analytics-data==0.12.0
protobuf==4.23.2
cachetools==5.3.1
